# Conflicts:
#	runtime/src/main/java/me/shedaniel/rei/impl/client/ClientHelperImpl.java
#	runtime/src/main/java/me/shedaniel/rei/impl/client/config/ConfigManagerImpl.java
#	runtime/src/main/java/me/shedaniel/rei/impl/client/config/ConfigObjectImpl.java
#	runtime/src/main/java/me/shedaniel/rei/impl/client/gui/ScreenOverlayImpl.java
#	runtime/src/main/java/me/shedaniel/rei/impl/client/gui/screen/DelegateScreen.java
#	runtime/src/main/java/me/shedaniel/rei/impl/client/gui/screen/TransformingScreen.java
#	settings.gradle

+++ default-plugin/src/main/java/me/shedaniel/rei/plugin/client/DefaultClientPlugin.java
@@ -26,11 +26,11 @@ package me.shedaniel.rei.plugin.client;
 import com.google.common.collect.Iterators;
 import com.google.common.collect.Lists;
 import com.google.common.collect.Sets;
+import dev.architectury.injectables.annotations.ExpectPlatform;
+import dev.architectury.injectables.annotations.PlatformOnly;
 import it.unimi.dsi.fastutil.objects.Object2FloatMap;
 import it.unimi.dsi.fastutil.objects.ReferenceOpenHashSet;
 import it.unimi.dsi.fastutil.objects.ReferenceSet;
-import me.shedaniel.architectury.annotations.ExpectPlatform;
-import me.shedaniel.architectury.annotations.PlatformOnly;
 import me.shedaniel.architectury.platform.Platform;
 import me.shedaniel.math.Rectangle;
 import me.shedaniel.rei.api.client.favorites.FavoriteEntry;
@@ -353,7 +353,7 @@ public class DefaultClientPlugin implements REIClientPlugin, BuiltinClientPlugin
     }
     
     @Override
-    public int getPriority() {
+    public double getPriority() {
         return -100;
     }
